Añadidos mensajes de estado y corrección de errores

+++ src/api/routes.py
@@ -17,55 +17,59 @@ def signup():
     if request.method == "GET":
         username = request.json.get("username", None)
         if not username:
-            return jsonify({"msg": "An username must be entered."})
+            return jsonify({"msg": "An username must be entered."}), 404
         user = User.query.filter_by(email = username).first()
         if user:
-            return jsonify({"info": user.serialize()})
+            return jsonify({"info": user.serialize()}), 200
         else:
-            return jsonify({"msg": "User doesn't exist."})
+            return jsonify({"msg": "User doesn't exist."}), 404
     
     if request.method == "POST":
         username = request.json.get("username", None)
         password = request.json.get("password", None)
         if not username or not password:
-            return jsonify({"msg": "An username and password must be entered."})
+            return jsonify({"msg": "An username and password must be entered."}), 400
         user = User(email = username,
                     password = password)
         db.session.add(user)
         db.session.commit()
-        return jsonify({"added": username})
+        return jsonify({"added": username}), 200
     
     if request.method == "PUT":
         username = request.json.get("username", None)
         password = request.json.get("password", None)
         if not username or not password:
-            return jsonify({"msg": "An username and password must be entered. Only the password will change."})
+            return jsonify({"msg": "An username and password must be entered. Only the password will change."}), 400
         user = User.query.filter_by(email = username).first()
         user.password = password
         db.session.commit()
-        return jsonify({"msg": "Password has been updated"})
+        return jsonify({"msg": "Password has been updated"}), 200
 
     if request.method == "DELETE":
         username = request.json.get("username", None)
         password = request.json.get("password", None)
         if not username or not password:
-            return jsonify({"msg": "An username and password must be entered."})
+            return jsonify({"msg": "An username and password must be entered."}), 400
         user = User.query.filter_by(email = username).first()
         if not user:
             return jsonify({"msg": "User not found"}), 404
         if user.password == password:
             db.session.delete(user)
             db.session.commit()
-            return jsonify({"removed": username})
+            return jsonify({"removed": username}), 200
         else:
-            return jsonify({"msg": "Password is wrong."})
+            return jsonify({"msg": "Password is wrong."}), 400
 
 # About tutors
 @api.route("/signup/tutor", methods=["POST", "PUT", "DELETE", "GET"])
 def signup_tutor():
     if request.method == "GET":
         username = request.json.get("username", None)
+        if not username or username == "":
+            return jsonify({"msg": "You must send a valid username"}), 400
         user = User.query.filter_by(email = username).first()
+        if not user:
+            return jsonify({"error": "User not found"}), 404
         tutor = Tutor.query.filter_by(user_id = user.id).first()
         if tutor:
             basic_response = tutor.serialize()
@@ -74,28 +78,35 @@ def signup_tutor():
             for kid in kids:
                 temp_kids = kid.serialize()
                 json_kids.append(temp_kids)
-            print("****************", json_kids)
-            print("****************", basic_response)
             return jsonify({"kids": json_kids,
-                            "msg": basic_response})
+                            "msg": basic_response}), 200
+        else:
+            return jsonify({"error": "Tutor not found"}), 400
+       
     
     if request.method == "POST":
         username = request.json.get("username", None)        
         if not username:
-            return jsonify({"msg": "An username must be entered."})
+            return jsonify({"msg": "An username must be entered."}), 400
         query_id = User.query.filter_by(email = username).first()
+        if not query_id:
+            return jsonify({"error": "Username not found"}), 404
         id = query_id.id
         tutor = Tutor(user_id = id)
         db.session.add(tutor)
         db.session.commit()
-        return jsonify({"added": username})
+        return jsonify({"added": username}), 200
 
     if request.method == "PUT":
         username = request.json.get("username", None)
         if not username:
-            return jsonify({"msg": "An username must be entered."})
+            return jsonify({"msg": "An username must be entered."}), 400
         user = User.query.filter_by(email = username).first().id
         tutor = Tutor.query.filter_by(user_id = user).first()
+        if not user:
+            return jsonify({"error": "Username not found"}), 404
+        if not tutor:
+            return jsonify({"error": "Tutor not found"}), 400
         serialize = Tutor.__table__.columns.keys()
         missing = []  
         for item in serialize:           
@@ -104,28 +115,30 @@ def signup_tutor():
                     missing.append(item)
         if missing != []:
             return jsonify({"msg": "Not all camps",
-                                "missing": missing})
+                                "missing": missing}), 404
         else:
             for item in serialize:
                 change = request.json.get(item,None)
                 if item != "id" and item != "user_id" and item != "children" and change != "":
                     tutor.query.update({item: change})
             db.session.commit()
-        return jsonify({"Changes": request.json})
+        return jsonify({"Changes": request.json}), 200
 
     if request.method == "DELETE":
         username = request.json.get("username", None)
         password = request.json.get("password", None)
         if not username or not password:
-            return jsonify({"msg": "An username and password must be entered."})
+            return jsonify({"msg": "An username and password must be entered."}), 400
         user = User.query.filter_by(email = username).first()
         if not user:
             return jsonify({"msg": "User not found"}), 404
         if user.password == password:
             tutor = Tutor.query.filter_by(user_id = user.id).first()
+            if not tutor:
+                return jsonify({"error": "Tutor not found"}), 400
             db.session.delete(tutor)
             db.session.commit()
-            return jsonify({"removed": username + " as tutor"})
+            return jsonify({"removed": username + " as tutor"}), 200
         else:
             return jsonify({"msg": "Password is wrong."})
 
@@ -134,9 +147,13 @@ def signup_child():
     if request.method == "POST":
         username = request.json.get("username", None)        
         if not username:
-            return jsonify({"msg": "An username must be entered."})
+            return jsonify({"msg": "An username must be entered."}), 400
         user_id = User.query.filter_by(email = username).first().id
         tutor_id = Tutor.query.filter_by(user_id = user_id).first().id
+        if not user_id:
+            return jsonify({"error": "Username not found"}), 404
+        if not tutor_id:
+            return jsonify({"error": "Tutor not found"}), 400
         if tutor_id:
             required = Child.__table__.columns.keys()
             missing = []
@@ -146,7 +163,7 @@ def signup_child():
                     missing.append(item)                    
         if missing != []:
             return jsonify({"msg": "Not all camps",
-                            "missing": missing})
+                            "missing": missing}), 404
         child = Child(name = request.json.get("name", None),
                       lastname = request.json.get("lastname", None),
                       parent = tutor_id,
@@ -158,13 +175,13 @@ def signup_child():
         tutor = Tutor.query.filter_by(user_id = user_id).first()
         tutor.children.append(child)
         db.session.commit()
-        return jsonify({"added": username})
+        return jsonify({"added": username}), 200
 
     if request.method == "PUT":
         child_id = request.json.get("child_id", None)
         checked_id = Child.query.filter_by(id = child_id).first()
         if not child_id or not checked_id:
-            return jsonify({"msg": "You need to send a valid child_id."})
+            return jsonify({"msg": "You need to send a valid child_id."}), 400
         required = Child.__table__.columns.keys()
         missing = []  
         for item in required:           
@@ -173,28 +190,26 @@ def signup_child():
                     missing.append(item)
         if missing != []:
             return jsonify({"msg": "Not all camps",
-                                "missing": missing})
+                                "missing": missing}), 404
         else:
             for item in required:
                 change = request.json.get(item,None)
                 if item != "id" and item != "parent" and change != "":
                     checked_id.query.update({item: change})
             db.session.commit()
-        return jsonify({"changes": request.json})
+        return jsonify({"changes": request.json}), 200
 
     if request.method == "DELETE":
         child_id = request.json.get("child_id", None)
         checked_id = Child.query.filter_by(id = child_id).first()
-        test = Child.query.filter_by(name = "Pepe").first()
         tutor = Tutor.query.filter_by(id = test.parent).first()
         if not child_id or not checked_id:
-            return jsonify({"msg": "You need to send a valid child_id.",
-                            "test": test.id})
+            return jsonify({"msg": "You need to send a valid child_id."}), 404
         child_name = checked_id.name + " " + checked_id.lastname
         tutor.children.remove(checked_id)
         db.session.delete(checked_id)
         db.session.commit()
-        return jsonify({"deleted": child_name})
+        return jsonify({"deleted": child_name}), 200
     
 
 @api.route("signup/advertiser", methods=["GET", "POST", "PUT", "DELETE"])
@@ -211,7 +226,7 @@ def advertiser():
         username = request.json.get("username", None)
         user = User.query.filter_by(email = username).first()
         if not username or not user:
-            return jsonify({"msg": "You must send a valid username"})
+            return jsonify({"msg": "You must send a valid username"}), 400
         required = ["name", "lastname", "contact", "company"]
         missing = []
         for item in required:           
@@ -219,7 +234,7 @@ def advertiser():
                 missing.append(item)                    
         if missing != []:
             return jsonify({"msg": "Not all camps",
-                            "missing": missing})
+                            "missing": missing}), 404
         advertiser = Advertiser(name = request.json.get("name", None),
                       lastname = request.json.get("lastname", None),
                       contact = request.json.get("contact", None),
@@ -227,14 +242,16 @@ def advertiser():
                       user_id = user.id)  
         db.session.add(advertiser)
         db.session.commit()
-        return jsonify({"created": request.json.get("name", None) + " " + request.json.get("lastname",None) })
+        return jsonify({"created": request.json.get("name", None) + " " + request.json.get("lastname",None) }), 200
 
     if request.method == "PUT":
         username = request.json.get("username", None)
         if not username:
-            return jsonify({"msg": "An username must be entered."})
+            return jsonify({"msg": "An username must be entered."}), 400
         user = User.query.filter_by(email = username).first().id
         advert = Advertiser.query.filter_by(user_id = user).first()
+        if not user or not advert:
+            return jsonify({"msg": "You must send a valid username"}), 404
         required = Advertiser.__table__.columns.keys()
         missing = []  
         for item in required:           
@@ -243,20 +260,20 @@ def advertiser():
                     missing.append(item)
         if missing != []:
             return jsonify({"msg": "Not all camps",
-                                "missing": missing})
+                                "missing": missing}), 404
         else:
             for item in required:
                 change = request.json.get(item,None)
                 if item != "id" and item != "user_id" and change != "":
                     advert.query.update({item: change})
             db.session.commit()
-        return jsonify({"Changes": request.json})
+        return jsonify({"Changes": request.json}), 200
 
     if request.method == "DELETE":
         username = request.json.get("username", None)
         password = request.json.get("password", None)
         if not username or not password:
-            return jsonify({"msg": "An username and password must be entered."})
+            return jsonify({"msg": "An username and password must be entered."}), 400
         user = User.query.filter_by(email = username).first()
         if not user:
             return jsonify({"msg": "User not found"}), 404
@@ -264,9 +281,9 @@ def advertiser():
             adv = Advertiser.query.filter_by(user_id = user.id).first()
             db.session.delete(adv)
             db.session.commit()
-            return jsonify({"removed": username + " as advertiser"})
+            return jsonify({"removed": username + " as advertiser"}), 200
         else:
-            return jsonify({"msg": "Password is wrong."}) 
+            return jsonify({"msg": "Password is wrong."}), 400
 
 
 @api.route("/event", methods=["GET", "POST", "PUT", "DELETE"])   
@@ -276,7 +293,12 @@ def event():
         event = Event.query.filter_by(id = event_id).first()
         if not event:
             return jsonify({"msg": "Event not found"}), 404
-        return jsonify({"info": event.serialize()})
+        participation = event.participants
+        response = []
+        for kid in participation:
+            response.append(kid.serialize())
+        return jsonify({"info": event.serialize(),
+                        "participants": response}), 200
 
     if request.method == "POST":
         username = request.json.get("username", None)
@@ -284,7 +306,7 @@ def event():
         adver = Advertiser.query.filter_by(user_id = user.id).first()
         required = Event.__table__.columns.keys()
         if not user or not adver:
-            return jsonify({"msg": "You must send a valid username that is an advertiser"})
+            return jsonify({"msg": "You must send a valid username that is an advertiser"}), 400
         missing = []
         for item in required:
             if item not in request.json and item != "id" and item != "id_advertiser" and item != score and item != score_amount and item != score_sum and item != done and item != participants:
@@ -308,23 +330,27 @@ def event():
                       others = request.json.get("others", None))
         db.session.add(event)   
         db.session.commit()
-        return jsonify({"msg": "Event created"})  
+        return jsonify({"msg": "Event created"}), 200 
 
     if request.method == "PUT":
         event_id = request.json.get("event_id", None)
         event = Event.request.json.get("event", None).first()
+        if not event:
+            return jsonify({"msg": "Event id not found"}), 404
         possible = Event.__table__.columns.keys()
         for item in request.json:
             if item in possible and item != "event_id":
                 event.query.update({item: change})
             elif item not in possible and item != "event_id":
-                return jsonify({"error": item + "can't be changed"})
+                return jsonify({"error": item + "can't be changed"}), 400
         db.session.commit()
-        return jsonify({"msg": "Event modified"})
+        return jsonify({"msg": "Event modified"}), 200
 
     if request.method == "DELETE":
         event_id = request.json.get("event_id", None)
         event = Event.request.json.get("event", None).first()
+        if not event:
+           return jsonify({"msg": "Event id not found"}), 404 
         db.session.delete(event)
         db.session.commit()
         return jsonify({"msg": "Event deleted"})
@@ -335,15 +361,22 @@ def login():
     username = request.json.get("username", None)
     password = request.json.get("password", None)
     user = User.query.filter_by(email = username).first()
-    if user:
+    if not username or not password:
+        return jsonify({"msg": "You need to send username and password"}), 400
+    if not user:
+        return jsonify({"msg": "User doesn't exist"}), 404
+    if user and user.password == password:
         access_token = create_access_token(identity=username)
-        return jsonify(access_token=access_token)
+        return jsonify(access_token=access_token), 200
     else:
-        return jsonify({"msg": "Error"})
+        return jsonify({"msg": "Error. Password is wrong."}), 400
 
 @api.route("/protected", methods=["GET"])
 @jwt_required()
 def protected():
     # Access the identity of the current user with get_jwt_identity
     current_user = get_jwt_identity()
-    return jsonify(logged_in_as=current_user), 200
+    if current_user:
+        return jsonify(logged_in_as=current_user), 200
+    else:
+        return jsonify({"msg": "Not authorized."}), 400
